third_party/WebKit/LayoutTests/inspector-enabled/sources/debugger/script-formatter-console.html

<html>
<head>
<script src="../../../http/tests/inspector/inspector-test.js"></script>
<script src="../../../http/tests/inspector/console-test.js"></script>
<script src="../../../http/tests/inspector/debugger-test.js"></script>
<script>

function nonFormattedFunction() { var  i = 2 + 2; var a = 4; return a + i; }

function f1()
{
    console.error("Test message.");
}

function onload()
{
    if (window.testRunner) {
        testRunner.waitUntilDone();
        testRunner.showWebInspector();
    }
    runTest();
}

var test = function()
{
    var panel = UI.panels.sources;
    var sourceFrame;
    var scriptFormatter;

    SourcesTestRunner.runDebuggerTestSuite([
        function testSetup(next)
        {
            SourcesTestRunner.scriptFormatter().then(function(sf) {
                scriptFormatter = sf;
                next();
            });
        },

        function testConsoleMessagesForFormattedScripts(next)
        {
            SourcesTestRunner.showScriptSource("script-formatter-console.html", didShowScriptSource);

            function didShowScriptSource(frame)
            {
                sourceFrame = frame;
                TestRunner.evaluateInPage("f1()", didEvaluate);
            }

            function didEvaluate()
            {
                dumpConsoleMessageURLs();
                TestRunner.addResult("Pre-format row message list:");
                TestRunner.addResult(JSON.stringify(sourceFrame._rowMessageBuckets.keysArray()));
                var name = panel.visibleView.uiSourceCode().name();
                scriptFormatter._toggleFormatScriptSource();
                SourcesTestRunner.showScriptSource(name + ":formatted", uiSourceCodeScriptFormatted);
            }

            function uiSourceCodeScriptFormatted()
            {
                dumpConsoleMessageURLs();
                TestRunner.addResult("Post-format row message list:");
                var formattedSourceFrame = panel.visibleView;
                TestRunner.addResult(JSON.stringify(formattedSourceFrame._rowMessageBuckets.keysArray()));
                next();
            }
        }
    ]);

    function dumpConsoleMessageURLs()
    {
        var messages = Console.ConsoleView.instance()._visibleViewMessages;
        for (var i = 0; i < messages.length; ++i) {
            var element = messages[i].toMessageElement();
            var anchor = element.querySelector(".console-message-anchor");
            TestRunner.addResult(anchor.textContent);
        }
    }
}

</script>

</head>

<body onload="onload()">
<p>Tests that the script formatting changes console line numbers.
</p>

</body>
</html>
